Provision and test NFS storage using the external subdir provider.
Shockingly (to me) this worked the first time.

The NFS server is just a QNAP hanging off the network; there's a share
that's ACL'd to allow only the cluster IPs to hit it.

Once the test nginx deployment is running, with the LB service giving
it a nice VIP, and the PV mounted, I can create arbitrary files in the
NFS subdirectory, and immediately access them via the nginx server.